logica tranferencia entre cuentas
aun falta el tema de validar pin de transaccion pero eso es un caso aparte ya que solo va a ser un metodo que va a llamar el GUITransferencia para validar el pin  con el de la cuenta segun la base de datos antes de realizar la tranferencia

+++ src/webbanking/operaciones/Deposito.java
@@ -13,10 +13,10 @@ import webbanking.operaciones.interfaces.DepositoInterface;
  */
 public class Deposito extends Operacion implements DepositoInterface {
     private double Deposito;
-    private double CDestino;
+    private long CDestino;
     private BaseDatos Basedatos=new BaseDatos();
     
-    public Deposito(double IdCuenta,double deposito){
+    public Deposito(long IdCuenta,double deposito){
         this.Deposito=deposito;
         this.CDestino=IdCuenta;
     }
